requirements.txt
# Dependencias principales
requests==2.31.0
python-dotenv==1.0.0

# Email y notificaciones
aiosmtplib==3.0.1
aiohttp==3.9.1

# XML parsing
xmltodict==0.13.0

# Google APIs
google-auth==2.35.0
google-api-python-client==2.156.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Utilidades
python-dateutil==2.8.2
